requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
barcode==1.0.2
Brotli==1.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==40.0.2
defusedxml==0.8.0rc2
Django==4.2
django-allauth==0.58.2
django-autoslug==1.9.9
django-bootstrap-datepicker-plus==5.0.4
django-ckeditor==6.7.0
django-compression-middleware==0.5.0
django-coreapi==2.3.0
django-crispy-forms==2.0
django-filter==23.1
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-silk==5.0.4
django-taggit==5.0.1
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.5
facebook-sdk==3.1.0
frozenlist==1.4.0
googlemaps==4.10.0
gprof2dot==2022.7.29
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
infobip-api-python-client==3.0.3
itypes==1.2.0
Jinja2==3.1.2
Levenshtein==0.20.9
Markdown==3.4.3
MarkupSafe==2.1.2
multidict==6.0.4
num2words==0.5.13
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pypng==0.20220715.0
python-barcode==0.14.0
python-dateutil==2.8.2
python-Levenshtein==0.20.9
python3-openid==3.2.0
pytz==2023.3
qrcode==7.4.2
rapidfuzz==2.15.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
sentry-sdk==1.34.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
social-auth-core==4.5.0
sqlparse==0.4.3
twilio==8.10.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.6.0
yarl==1.9.2
zipp==3.17.0
zstandard==0.22.0
